+++ frontend/src/components/Trade/TradeDetailDeal.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { TopBar, TradeTopBar } from "../common/Navigator/navigator";
+import { DealTopBar, TopBar, TradeTopBar } from "../common/Navigator/navigator";
 import Coli from "/src/assets/images/브로콜리.png";
 import { LayoutInnerBox, LayoutMainBox } from "../common/Layout/Box";
 import { BottomButton } from "../common/Button/LargeButton";
@@ -11,14 +11,20 @@ import Like from "/src/assets/svg/like.svg";
 import { useState } from "react";
 import Tree from "/src/assets/svg/diarytree.svg";
 import { useMutation, useQuery } from "@tanstack/react-query";
-import { deletePost, getTradeDetail, useLike } from "../../apis/TradeApi";
+import {
+  deletePost,
+  getTradeDetail,
+  useDealFinish,
+  useLike,
+} from "../../apis/TradeApi";
 import { useParams } from "react-router-dom";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css";
 import "swiper/css/navigation";
 import SwiperCore, { Navigation, Pagination } from "swiper/modules";
-import { userCheck } from "../../apis/UserApi";
-import TradePostLayout from "../common/Layout/TradePostLayout";
+import { userCheck, userIndividualCheck } from "../../apis/UserApi";
+import React, { ChangeEvent, useEffect } from "react";
+import { getUsersInfo } from "../../apis/MyPage";
 interface ImageResponse {
   imgStoreUrl: string;
 }
@@ -36,7 +42,7 @@ const ImgBox = styled.img`
 `;
 const InfoBox = styled.div`
   width: 100%;
-  height: 2.125rem;
+  height: 3.125rem;
   display: flex;
   justify-content: space-between;
   gap: 8.81rem;
@@ -47,7 +53,7 @@ const Profile = styled.div`
   display: flex;
   flex-direction: row;
   align-items: center;
-  gap: 0.2rem;
+  gap: 0.6rem;
 `;
 const Name = styled.div`
   width: auto;
@@ -112,7 +118,7 @@ const NavigateText = styled.div`
 `;
 const ExplainText = styled.div`
   color: #000;
-  font-size: 0.875rem;
+  font-size: 1.2rem;
   font-weight: 400;
   width: 100%;
   height: auto;
@@ -125,6 +131,7 @@ const SwiperContainer = styled.div`
 const Thumbnail = styled.img`
   width: 1.875rem;
   height: 1.875rem;
+  border-radius: 50%;
 `;
 const PriceBox = styled.div`
   width: auto;
@@ -135,6 +142,7 @@ const PriceBox = styled.div`
   font-size: 1.25rem;
   font-weight: bold;
 `;
+
 const StateIcon = styled.div<Icon & { children?: React.ReactNode }>`
   width: ${(props) => `${props.width}rem`};
   height: ${(props) => `${props.height}rem`};
@@ -146,25 +154,85 @@ const StateIcon = styled.div<Icon & { children?: React.ReactNode }>`
   align-items: center;
   font-size: 0.5625rem; /* 텍스트 크기 */
 `;
+
+const DoneBtn = styled.button`
+  width: 3.9375rem;
+  height: 1.75rem;
+  border-radius: 0.625rem;
+  background: var(--a-0-d-8-b-3, #2a7f00);
+  text-align: center;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
 const TradeDetailDeal = () => {
   const navigate = useNavigate();
+  const [tradeDetail, setTradeDetail] = useState(null); // 상태를 초기화합니다.
+  const [authorId, setAuthorId] = useState(null); // 상태를 초기화합니다.
+  const [loginid, setLoginId] = useState(null); // 상태를 초기화합니다.
+
+  const { postId } = useParams<{ postId: string }>();
+
+  // useEffect(() => {
+  //   const fetchData = async () => {
+  //     try {
+  //       const accessToken = sessionStorage.getItem("accessToken"); // 세션 스토리지에서 accessToken을 가져옵니다.
+
+  //       const userinfo = await getUsersInfo();
+
+  //       console.log("userinfo" + JSON.stringify(userinfo));
+
+  //       console.log("userinfo" + userinfo.data.data_body.id);
+  //       setLoginId(userinfo.data.data_body.id);
+
+  //       console.log("postid:" + postId);
+  //       const postIdNumber = postId ? parseInt(postId) : undefined;
+
+  //       if (accessToken !== null) {
+  //         // accessToken이 null이 아닌 경우에만 실행합니다.
+
+  //         if (postIdNumber) {
+  //           const data = await getTradeDetail(accessToken, postIdNumber); // 비동기 함수를 호출합니다.
+  //           console.log("dettttttttttttt", JSON.stringify(data)); // data 객체를 직렬화하여 출력합니다.
+  //           console.log("dettttttttttttt", data.exArticleResponse.userId); // data 객체를 직렬화하여 출력합니다.
+  //           setAuthorId(data.exArticleResponse.userId);
+  // g        }
+
+  //         setTradeDetail(data); // 데이터를 상태에 저장합니다.
+  //       }
+  //     } catch (error) {
+  //       console.error("에러 발생:", error);
+  //     }
+  //   };
+
+  // fetchData(); // 함수를 호출합니다.
+  // }, []); // 빈 배열을 전달하여 컴포넌트가 처음 렌더링될 때 한 번만 호출되도록 설정합니다.
 
   const [like, setLike] = useState<boolean>(false);
   const handleLike = () => {
     setLike(!like);
   };
 
-  const { mutate: handleLikeClick } = useLike();
-  const { postId } = useParams<{ postId: string }>();
   const postNumber = Number(postId);
   const accessToken = sessionStorage.getItem("accessToken");
+  const [diary, setDiary] = useState<number>();
+
   const { isLoading, data, error } = useQuery({
     queryKey: ["tradeDetail", postNumber],
     queryFn: accessToken
       ? () => getTradeDetail(accessToken, postNumber)
       : undefined,
   });
+  console.log("디테일 데이터", data);
+  // console.log("디테일 데이터", data?.packDiaryResponse);
 
+  // setDiary(data?.packDiaryResponse)
+
+  // console.log("디테일 데이터", data?.dealResponse);
+
+  // console.log("디테일 데이터", data.data.packDiaryResponse);
+
+  const { mutate: handleLikeClick } = useLike({ queryKeys: ["tradeDetail"] });
   const {
     isLoading: isLoadingUserDetail,
     data: userData,
@@ -174,10 +242,10 @@ const TradeDetailDeal = () => {
     queryFn: accessToken ? () => userCheck(accessToken) : undefined,
   });
   const userId = userData?.id;
-
+  console.log(data, "데이터입니다");
   const DiaryId = data?.packDiaryResponse?.packDiaryId;
   const handleDiary = (DiaryId: number) => {
-    navigate(`/diary/${DiaryId}`);
+    navigate(`/crop/${DiaryId}/otherview`);
     console.log("나 눌리고 있어!!!", 111);
   };
   const formatDateAndTime = (dateString: string) => {
@@ -187,19 +255,24 @@ const TradeDetailDeal = () => {
     return `${date} ${hours}:${minutes}:${seconds}`;
   };
 
-  console.log(
-    "데이터 id",
-    data?.exArticleResponse?.userId,
-    "유저 id",
-    userData?.id
-  );
-  const BtnClick = (postId: number) => {
-    if (data?.exArticleResponse?.userId === userData?.id) {
-      navigate(`/trade/${postId}/seller`);
-    } else {
-      navigate(`/trade/${postId}/buyer`);
-    }
-    // console.log("저를 클릭했나요?");
+  const {
+    isLoading: isIndividualUserDetail,
+    data: IndividualUserData,
+    error: IndividualUserDetailError,
+  } = useQuery({
+    queryKey: ["individualUserDetail"],
+    queryFn: () =>
+      userIndividualCheck(
+        accessToken as string,
+        data?.exArticleResponse.userId
+      ),
+    enabled: !!accessToken && !!data?.exArticleResponse.userId, // 여기에 조건 추가
+  });
+  const handleSellerClick = (postId: number) => {
+    navigate(`/trade/${postId}/seller`);
+  };
+  const handleBuyerClick = (postId: number) => {
+    navigate(`/trade/${postId}/buyer`);
   };
 
   const handleEdit = () => {
@@ -212,10 +285,11 @@ const TradeDetailDeal = () => {
         ex_article_location: data?.exArticleResponse?.exLocation,
         packdiaryid: data?.packDiaryResponse?.packDiaryId.toString(),
         deal_cur_price: data?.dealResponse?.price.toString(),
-        imageResponse: data?.imageResponses,
+        imageResponse: data?.exArticleResponse.imageResponses,
         postId: data?.exArticleResponse.exArticleId,
       },
     });
+    // console.log('gettradedetail'+response.data.data_body.exArticleResponse.imageResponses[0].imgStoreUrl)
   };
   const { mutate: deleteMutation } = useMutation({
     mutationFn: deletePost,
@@ -227,15 +301,53 @@ const TradeDetailDeal = () => {
       console.log(err);
     },
   });
+  // const handleDeleteConfirmation = (postId: number) => {
+  //   const isConfirmed = window.confirm("삭제하시겠습니까?"); // 사용자에게 삭제 확인 요청
+  //   if (isConfirmed) {
+  //     // 사용자가 '확인'을 클릭한 경우
+  //     deleteMutation(postId); // 삭제 함수 실행
+  //   }
+  // };
+  console.log("머임", typeof data?.exArticleResponse.exArticleId);
+  // 거래 종료
+  // const handleFinishClick = () => {
+  //   window.confirm("거래를 종료하시겠습니까?");
+  //   finishClick(data?.exArticleResponse.exArticleId);
+  //   navigate("/trade");
+  // };
+
   return (
     <>
-      <TradeTopBar
-        title="작물거래"
+      {/* {data?.exArticleResponse.userId === data?.userResponse.id ? (
+        <TradeTopBar
+          title="작물거래"
+          showBack={true}
+          showEdit={true}
+          onEdit={handleEdit}
+          onDelete={() => {
+            handleDeleteConfirmation(data?.exArticleResponse.exArticleId);
+          }}
+        />
+      ) : (
+        <TradeTopBar
+          title="작물거래"
+          showBack={true}
+          showEdit={false}
+          // onEdit={handleEdit}
+          // onDelete={() => {
+          //   deleteMutation(data?.exArticleResponse.exArticleId);
+          // }}
+        />
+      )} */}
+      <DealTopBar
         showBack={true}
+        title="작물거래"
         showEdit={true}
-        onEdit={handleEdit}
         onDelete={() => {
-          deleteMutation(data?.exArticleResponse.exArticleId);
+          const isConfirmed = window.confirm("정말로 삭제하시겠습니까?");
+          if (isConfirmed) {
+            deleteMutation(data?.exArticleResponse.exArticleId);
+          }
         }}
       />
       <LayoutMainBox>
@@ -256,15 +368,16 @@ const TradeDetailDeal = () => {
         <LayoutInnerBox>
           <InfoBox>
             <Profile>
-              <Thumbnail src={data?.userResponse?.thumbnail} alt="profile" />
+              <Thumbnail src={IndividualUserData?.thumbnail} alt="profile" />
               <Name>
-                <NameText>{data?.userResponse.nickname}</NameText>
+                <NameText>{IndividualUserData?.nickname}</NameText>
                 <ClassesText>
-                  {data?.userResponse.rank}
+                  {IndividualUserData?.rank}
                   {/* <img src={Sprout} alt="Sprout" /> */}
                 </ClassesText>
               </Name>
             </Profile>
+
             <Date>{formatDateAndTime(data?.exArticleResponse.time)}</Date>
           </InfoBox>
           <TitleBox>
@@ -291,21 +404,35 @@ const TradeDetailDeal = () => {
             </PriceBox>
 
             <DiaryBox>
-              <img src={Tree} alt="tree" />
-              <NavigateText
-                onClick={() => {
-                  DiaryId ? handleDiary(Number(DiaryId)) : null;
-                }}
-              >
-                작물일지 이동하기
-              </NavigateText>
+              {data?.packDiaryResponse ? (
+                <>
+                  <img src={Tree} alt="tree" />
+                  <NavigateText
+                    onClick={() => {
+                      DiaryId ? handleDiary(Number(DiaryId)) : null;
+                    }}
+                  >
+                    작물일지 이동하기
+                  </NavigateText>
+                </>
+              ) : null}
             </DiaryBox>
             <ExplainText>{data?.exArticleResponse.content}</ExplainText>
           </TitleBox>
         </LayoutInnerBox>
+
+        {/* 구매자면 가격 제안조회 고 판매자면 가격 제안하기. */}
         <BottomButton
-          text="가격 제안하기"
-          onClick={() => BtnClick(postNumber)}
+          text={
+            data?.userResponse.id === data?.exArticleResponse.userId
+              ? "제안목록 확인"
+              : "제안하기"
+          }
+          onClick={() =>
+            data?.userResponse.id === data?.exArticleResponse.userId
+              ? handleSellerClick(data?.exArticleResponse.exArticleId)
+              : handleBuyerClick(data?.exArticleResponse.exArticleId)
+          }
         />
       </LayoutMainBox>
     </>
