+++ frontend/src/components/Trade/TradeDetailGeneral.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { TopBar, TradeTopBar } from "../common/Navigator/navigator";
+import { DealTopBar, TopBar, TradeTopBar } from "../common/Navigator/navigator";
 import Coli from "/src/assets/images/브로콜리.png";
 import { LayoutInnerBox, LayoutMainBox } from "../common/Layout/Box";
 import { BottomButton } from "../common/Button/LargeButton";
@@ -11,18 +11,27 @@ import Like from "/src/assets/svg/like.svg";
 import { useState } from "react";
 import Tree from "/src/assets/svg/diarytree.svg";
 import { useMutation, useQuery } from "@tanstack/react-query";
-import { deletePost, getTradeDetail, useLike } from "../../apis/TradeApi";
+import {
+  createChatRoom,
+  deletePost,
+  getTradeDetail,
+  useLike,
+} from "../../apis/TradeApi";
 import { useParams } from "react-router-dom";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css";
 import "swiper/css/navigation";
 import SwiperCore, { Navigation, Pagination } from "swiper/modules";
-import { userCheck } from "../../apis/UserApi";
+import { userCheck, userIndividualCheck } from "../../apis/UserApi";
 import TradePostLayout from "../common/Layout/TradePostLayout";
+import { useAtom } from "jotai";
+import { likeAtom } from "../../stores/trade";
 interface ImageResponse {
   imgStoreUrl: string;
 }
-
+interface ClickLike {
+  onClick: (e: React.MouseEvent<HTMLDivElement>) => void;
+}
 const ImgBox = styled.img`
   width: 100%;
   height: 15.5625rem;
@@ -41,7 +50,7 @@ const Profile = styled.div`
   display: flex;
   flex-direction: row;
   align-items: center;
-  gap: 0.2rem;
+  gap: 0.6rem;
 `;
 const Name = styled.div`
   width: auto;
@@ -119,19 +128,29 @@ const SwiperContainer = styled.div`
 const Thumbnail = styled.img`
   width: 1.875rem;
   height: 1.875rem;
+  border-radius: 50%;
+`;
+const LikeBox = styled.div<ClickLike>`
+  width: 1.25rem;
+  height: 1.25rem;
+  display: flex;
+  position: absolute;
+  bottom: 0.38rem;
+  right: 0.38rem;
+  z-index: 3;
+  cursor: pointer;
 `;
 
 const TradeGeneralDetail = () => {
   const navigate = useNavigate();
-  const BtnClick = () => {
-    navigate("/trade/chat");
-  };
+  // const BtnClick = (postId: number) => {
+  //   navigate(`/trade/${postId}/chat`);
+  // };
   const [like, setLike] = useState<boolean>(false);
   const handleLike = () => {
     setLike(!like);
   };
 
-  const { mutate: handleLikeClick } = useLike();
   const { postId } = useParams<{ postId?: string }>();
   const postNumber = Number(postId);
   const accessToken = sessionStorage.getItem("accessToken");
@@ -141,6 +160,7 @@ const TradeGeneralDetail = () => {
       ? () => getTradeDetail(accessToken, postNumber)
       : undefined,
   });
+  const { mutate: handleLikeClick } = useLike({ queryKeys: ["tradeDetail"] });
   const {
     isLoading: isLoadingUserDetail,
     data: userData,
@@ -149,11 +169,28 @@ const TradeGeneralDetail = () => {
     queryKey: ["userDetail"],
     queryFn: accessToken ? () => userCheck(accessToken) : undefined,
   });
+
+  // 게시글 작성자의 썸네일이나 닉네임이 detail 조회하는 api에 들어있지 않아서 따로 조회해서 사용
+  const {
+    isLoading: isIndividualUserDetail,
+    data: IndividualUserData,
+    error: IndividualUserDetailError,
+  } = useQuery({
+    queryKey: ["individualUserDetail"],
+    queryFn: () =>
+      userIndividualCheck(
+        accessToken as string,
+        data?.exArticleResponse.userId
+      ),
+    enabled: !!accessToken && !!data?.exArticleResponse.userId, // 여기에 조건 추가
+  });
+
+  const { mutate: clickChat } = createChatRoom();
   const userId = userData?.id;
 
   const DiaryId = data?.packDiaryResponse?.packDiaryId;
   const handleDiary = (DiaryId: number) => {
-    navigate(`/diary/${DiaryId}`);
+    navigate(`/crop/${DiaryId}/otherview`);
     console.log("나 눌리고 있어!!!", 111);
   };
   const formatDateAndTime = (dateString: string) => {
@@ -187,15 +224,45 @@ const TradeGeneralDetail = () => {
       console.log(err);
     },
   });
+  const handleBuyerChatClick = () => {
+    clickChat(postNumber);
+  };
+  const handleSellerChatClick = () => {
+    navigate("/trade/chatroom");
+  };
+  console.log("data", data);
   return (
     <>
-      <TradeTopBar
-        title="작물거래"
+      {/* {data?.exArticleResponse.userId === data?.userResponse.id ? (
+        <TradeTopBar
+          title="작물거래"
+          showBack={true}
+          showEdit={true}
+          onEdit={handleEdit}
+          onDelete={() => {
+            deleteMutation(data?.exArticleResponse.exArticleId);
+          }}
+        />
+      ) : (
+        <TradeTopBar
+          title="작물거래"
+          showBack={true}
+          showEdit={false}
+          // onEdit={handleEdit}
+          onDelete={() => {
+            deleteMutation(data?.exArticleResponse.exArticleId);
+          }}
+        />
+      )} */}
+      <DealTopBar
         showBack={true}
+        title="작물거래"
         showEdit={true}
-        onEdit={handleEdit}
         onDelete={() => {
-          deleteMutation(data?.exArticleResponse.exArticleId);
+          const isConfirmed = window.confirm("정말로 삭제하시겠습니까?");
+          if (isConfirmed) {
+            deleteMutation(data?.exArticleResponse.exArticleId);
+          }
         }}
       />
       <LayoutMainBox>
@@ -216,11 +283,11 @@ const TradeGeneralDetail = () => {
         <LayoutInnerBox>
           <InfoBox>
             <Profile>
-              <Thumbnail src={data?.userResponse.thumbnail} alt="profile" />
+              <Thumbnail src={IndividualUserData?.thumbnail} alt="profile" />
               <Name>
-                <NameText>{data?.userResponse.nickname}</NameText>
+                <NameText>{IndividualUserData?.nickname}</NameText>
                 <ClassesText>
-                  {data?.userResponse.rank}
+                  {IndividualUserData?.rank}
                   {/* <img src={Sprout} alt="Sprout" /> */}
                 </ClassesText>
               </Name>
@@ -240,26 +307,34 @@ const TradeGeneralDetail = () => {
             </Title>
             <Price>{data?.transResponse.price}원</Price>
             <DiaryBox>
-              <img src={Tree} alt="tree" />
-              <NavigateText
-                onClick={() => {
-                  DiaryId ? handleDiary(Number(DiaryId)) : null;
-                }}
-              >
-                작물일지 이동하기
-              </NavigateText>
+              {data?.packDiaryResponse ? (
+                <>
+                  <img src={Tree} alt="tree" />
+                  <NavigateText
+                    onClick={() => {
+                      DiaryId ? handleDiary(Number(DiaryId)) : null;
+                    }}
+                  >
+                    작물일지 이동하기
+                  </NavigateText>
+                </>
+              ) : null}
             </DiaryBox>
-            <ExplainText>
-              심우석의 머리를 브로콜리에 비유하는 것은 그의 독특하고 특이한 헤어
-              스타일을 묘사하기 위한 창의적인 방법입니다. 이 비유는 특히 그의
-              머리카락이 풍성하고 볼륨감이 많으며, 위로 솟아 오른 모양이 마치
-              브로콜리의 녹색 송이와 유사하다는 점에서 온 것일 수 있습니다.
-              브로콜리의 작은 꽃송이들이 모여 있는 모양은, 심우석의 머리카락이
-              여러 방향으로 풍성하게 서 있는 것과 비슷하다고 할 수 있습니다.
-            </ExplainText>
+            <ExplainText>{data?.exArticleResponse.content}</ExplainText>
           </TitleBox>
         </LayoutInnerBox>
-        <BottomButton text="채팅하기" onClick={BtnClick} />
+        <BottomButton
+          text={
+            data?.userResponse.id === data?.exArticleResponse.userId
+              ? "채팅방으로 이동하기"
+              : "채팅하기"
+          }
+          onClick={
+            data?.userResponse.id === data?.exArticleResponse.userId
+              ? handleSellerChatClick
+              : handleBuyerChatClick
+          }
+        />
       </LayoutMainBox>
     </>
   );
